Move Parcel.findAllMatching to util/FindParcelsMatching
