+++ packages/parcels/src/parcel/ValueParcelMethods.js
@@ -4,11 +4,7 @@ import type Parcel from './Parcel';
 import strip from '../parcelData/strip';
 import ChangeRequest from '../change/ChangeRequest';
 import ActionCreators from '../change/ActionCreators';
-import {containsWildcard, split} from '../modifiers/Matcher';
 
-import flatMap from 'unmutable/lib/flatMap';
-import shallowEquals from 'unmutable/lib/shallowEquals';
-import take from 'unmutable/lib/take';
 import pipeWith from 'unmutable/lib/util/pipeWith';
 
 export default (_this: Parcel): Object => ({
@@ -51,44 +47,6 @@ export default (_this: Parcel): Object => ({
 
     getInternalLocationShareData: (): Object => {
         return _this._treeshare.locationShare.get(_this.path());
-    },
-
-    findAllMatching: (match: string): Parcel[] => {
-        let matchParts = split(match);
-        let path = _this.path();
-
-        let baseMatches = pipeWith(
-            matchParts,
-            take(path.length),
-            shallowEquals(path)
-        );
-
-        if(!baseMatches) {
-            return [];
-        }
-
-        let get = (parcel: Parcel, matchParts: string[]): Parcel[] => {
-            let [matchPart, ...remainingMatchParts] = matchParts;
-
-            if(!matchPart) {
-                return [parcel];
-            }
-            if(!parcel.isParent()) {
-                return [];
-            }
-            if(containsWildcard(matchPart)) {
-                return pipeWith(
-                    parcel.toArray(pp => get(pp, remainingMatchParts)),
-                    flatMap(ii => ii)
-                );
-            }
-            if(!parcel.has(matchPart)) {
-                return [];
-            }
-            return get(parcel.get(matchPart), remainingMatchParts);
-        };
-
-        return get(_this, matchParts.slice(path.length));
     },
 
     // change methods
